Add LAEs and LBGs
This commit adds mass calculations and parameter file outputs for
* The Lyman Break Galaxies from Rotermund et al. (2021)
* The Lyman Alpha Emitters from Apostolovski et al.

This also adds the LAE and LBG data from Hill et al. (2021) except for LAE1. LAE1 is clearly a mis-identification because it is well-detected in the g,r,i optical bands but objects at redshift=4.3 need to ‘drop out’ of the g-band (Scott).