Change C1 gas fraction to 0.7
Originally we used a gas fraction of 0.4 for C1 on the rationale that the error in Hill+ 2021's stellar,
mass measurement is due to the foreground object, not due to lots of gas as in the other galaxies. We no longer think this is a good assumption, so now we use a gas fraction of 0.7 as in the other galaxies.

This commit also changes compute_galaxy_mass.ipynb so that the gas fractions are more explicitly set and the string describing the method is automatically generated based on the new fgas variables (rather than manually).