Change mass files to be fixed-width rather than comma-separated
This commit
* Replaces the .csv mass files with fixed-width .txt files so that they're easier to read.
* Updates the python code for this new format.

Note: compute_galaxy_mass.ipynb will now save the mass files with fewer decimal places. This results in a small rounding error – nevertheless, this should be inconsequential because this project does not require perfect precision, but aims to capture the character of the galaxies. The rounding error should be *much* smaller than 1%.

(The output files in generate/makegalaxy/ changed due to this rounding error.)