Include all galaxies within 105 kpc of centre
Previously we included all galaxies within 95 kpc; now this is 105 kpc. This adds LAE-ID72 to the galaxies under study.

Originally, LAE-ID72 was within 95 kpc of the centre and so was included, but refined mass estimates changed the centre of mass so as to put it just outside this radius. This commit increases the radius included to re-include it.