Switch to using Hill+2020 mass plus one sigma
Previously we used the Hill+2020 total mass reported. Now we use their upper bound (i.e. plus one sigma).

This results in a halo-to-stellar-mass ratio more in line with Behroozi+ so it seems like a reasonable "compromise", especially given the uncertain nature of the assumptions going into both the Hill+ and the Behroozi+ predictions.