Add chosen initial conditions for each galaxy, to be spliced.
This commit adds .hdf5 GIZMO snapshot files with the initial conditions that we will use for each galaxy. These result from simulating each galaxy in isolation. This commit also adds .pdf plots for each galaxy, showing the SFR while simulating in isolation, with arrows pointing to the chosen snapshots. The plots include the amount of gas mass lost and the change in gas fraction compared to the original makegalaxy output, in the legends. (The change is due to star formation while we simulate in isolation.)